mcp-python/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "addict>=2.4.0",
    "datasets==2.16.0",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain-modelscope-integration>=0.1.0",
    "modelscope>=1.28.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "simplejson>=3.20.1",
    "sortedcontainers>=2.4.0",
    "torch>=2.7.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
